Backend/requirements.txt
lxml==5.3.0
bs4==0.0.2
openai==1.42.0
faiss-cpu==1.8.0

langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.34
langchain_openai==0.1.22
langgraph

python-dotenv==1.0.0
python-multipart==0.0.6
### python --> 3.10.15 
streamlit==1.41.1
upstash-semantic-cache==1.0.0

httpx==0.27.0  # For the error pydantic.v1.error_wrappers.ValidationError: 1 validation error for ChatOpenAI __root__ Client.__init__() got an unexpected keyword argument 'proxies' (type=type_error)